Handle Algolia's configure prop not working
