+++ site/gatsby-site/src/components/discover/parseURL.js
@@ -1,0 +1,113 @@
+import { parse } from 'query-string';
+import { decodeQueryParams } from 'use-query-params';
+
+const removeUndefinedAttributes = (obj) => {
+  for (const attr in obj) {
+    if (obj[attr] === undefined) {
+      delete obj[attr];
+    }
+  }
+  return { ...obj };
+};
+
+const convertStringToArray = (obj) => {
+  const stringKeys = [
+    'source_domain',
+    'authors',
+    'submitters',
+    'incident_id',
+    'flag',
+    'classifications',
+    'is_incident_report',
+    'tags',
+    'language',
+  ];
+
+  let newObj = {};
+
+  for (const attr in obj) {
+    if (stringKeys.includes(attr) && obj[attr] !== undefined) {
+      if (obj[attr].indexOf('CSET') >= 0) {
+        // The facet separator is double pipe sign - "||"
+        newObj[attr] = obj[attr].split('||CSET').map((i) => 'CSET' + i);
+        newObj[attr][0] = newObj[attr][0].substr(4);
+      } else {
+        newObj[attr] = obj[attr].split('||');
+      }
+    } else {
+      if (obj[attr]) {
+        newObj[attr] = obj[attr];
+      }
+    }
+  }
+  return newObj;
+};
+
+const convertStringToRange = (query) => {
+  const rangeKeys = [
+    'epoch_incident_date_min',
+    'epoch_incident_date_max',
+    'epoch_date_published_min',
+    'epoch_date_published_max',
+  ];
+
+  let resultObj = {};
+
+  for (const attr in query) {
+    if (rangeKeys.includes(attr)) {
+      if (attr.includes('_min') && !isNaN(parseInt(query[attr]))) {
+        resultObj[attr.split('_min')[0]] = {
+          ...resultObj[attr.split('_min')[0]],
+          min: parseInt(query[attr]),
+        };
+      }
+
+      if (attr.includes('_max') && !isNaN(parseInt(query[attr]))) {
+        resultObj[attr.split('_max')[0]] = {
+          ...resultObj[attr.split('_max')[0]],
+          max: parseInt(query[attr]),
+        };
+      }
+    }
+  }
+
+  return resultObj;
+};
+
+const generateSearchState = ({ query }) => {
+  const cleanQuery = removeUndefinedAttributes(query);
+
+  const querySearch = cleanQuery.s || '';
+
+  delete cleanQuery.s;
+
+  return {
+    page: query.page,
+    query: querySearch,
+    refinementList: {
+      ...convertStringToArray(cleanQuery),
+    },
+    range: {
+      ...convertStringToRange(cleanQuery),
+    },
+    sortBy: cleanQuery.sortBy,
+
+    // TODO: https://github.com/algolia/instantsearch/issues/5892
+    __configure: {
+      hitsPerPage: 28,
+      distinct: true,
+    },
+  };
+};
+
+export default function ({ location, indexName, queryConfig }) {
+  const object = parse(location.search);
+
+  const query = decodeQueryParams(queryConfig, object);
+
+  const searchState = generateSearchState({ query });
+
+  console.log('holis parseurl', searchState);
+
+  return { [indexName]: searchState };
+}
